requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
install==1.3.4
joblib==1.1.0
mccabe==0.6.1
nltk==3.6.5
numpy==1.21.4
pycodestyle==2.8.0
pyflakes==2.4.0
pymystem3==0.2.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.11.2
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
zipp==3.6.0